refactor(schemas): renamed zod schemas files to legacy
