+++ apps/papra-server/src/modules/api-keys/api-keys.schemas.legacy.ts
