+++ apps/papra-server/src/modules/custom-properties/custom-properties.schemas.legacy.ts
