+++ apps/papra-server/src/modules/custom-properties/definitions/select/select.custom-property-definition.schemas.legacy.ts
