+++ apps/papra-server/src/modules/custom-properties/options/custom-properties-options.schemas.legacy.ts
