+++ apps/papra-server/src/modules/documents/documents.schemas.legacy.ts
