+++ apps/papra-server/src/modules/organizations/organization.schemas.legacy.ts
