+++ apps/papra-server/src/modules/tagging-rules/tagging-rules.schemas.legacy.ts
