+++ apps/papra-server/src/modules/users/users.schemas.legacy.ts
