+++ apps/papra-server/src/modules/webhooks/webhooks.schemas.legacy.ts
